test(gui): moved dialog_window fixture for create_dialog unit tests to conftest
